test: add mock for e2e tests and snapshot for regression forapi data gouv ess
