+++ clients/_test/clientApiDataGouvEss/_snapshots/ai-data-gouv-ess-800329849.json
@@ -1,0 +1,10 @@
+{
+  "args": [
+    "800329849"
+  ],
+  "result": {
+    "familleJuridique": "Associations",
+    "nom": "THEATRE STEPHANE GILDAS",
+    "region": "ILE DE FRANCE"
+  }
+}
