+++ clients/_test/clientApiDataGouvEss/index.test.ts
@@ -1,0 +1,14 @@
+import { clientEss } from '#clients/api-data-gouv/ess';
+import { Siren } from '#utils/helpers';
+import { expectClientToMatchSnapshot } from '../expect-client-to-match-snapshot';
+
+describe('clientAssociation', () => {
+  it('Should match snapshot', async () => {
+    await expectClientToMatchSnapshot({
+      client: clientEss,
+      args: ['800329849' as Siren],
+      snaphotFile: 'api-data-gouv-ess-800329849.json',
+      __dirname,
+    });
+  });
+});
