+++ cypress/fixtures/api-data-gouv-ess.json
@@ -1,0 +1,25 @@
+{
+  "data": [
+    {
+      "SIREN": "800329849",
+      "Nom ou raison sociale de l'entreprise": "THEATRE STEPHANE GILDAS",
+      "Famille juridique de l'entreprise": "Associations",
+      "Code postal": "75013",
+      "Libellé de la commune de l'établissement": "PARIS 13",
+      "Code du département de l'établissement": "75",
+      "Département de l'établissement": "PARIS",
+      "Région de l'établissement": "ILE DE FRANCE"
+    }
+  ],
+  "links": {
+    "profile": "https://tabular-api.data.gouv.fr/api/resources/57bc99ca-0432-4b46-8fcc-e76a35c9efaf/profile/",
+    "swagger": "https://tabular-api.data.gouv.fr/api/resources/57bc99ca-0432-4b46-8fcc-e76a35c9efaf/swagger/",
+    "next": null,
+    "prev": null
+  },
+  "meta": {
+    "page": 1,
+    "page_size": 20,
+    "total": 1
+  }
+}
