+++ cypress/mocks/handlers/api-data-gouv-ess.ts
@@ -1,0 +1,6 @@
+import { HttpResponse, HttpResponseResolver } from 'msw';
+import apiDataGouvEss from '../../fixtures/api-data-gouv-ess.json';
+
+export const apiDataGouvEssHandler: HttpResponseResolver = ({ request }) => {
+  return HttpResponse.json(apiDataGouvEss);
+};
